RecipeService.java
Greetings.
I tried running you code and it has a lot bugs. So I tried to make the necessary changes to it. So I made the following changes to line 54, 84. Since you were using a JSONObject, a try and catch block needs to be implemented. If you try running the code without it, it won't even comply.